[cms] feat: change proposaltoken to proposalname on invoice lineitem
This diff now allows users to select proposals by its name, instead of
its token.

Since getting proposals names require a batching request, and those
requests are bounded by 20 proposals for each request, those requests
have to be done more than once in case the token from the
invoice (details and diff), or
the proposal (new and edit actions) are not on the dropdown list.