Add docs/release-notes.
This diff creates a `docs/release-notes` directory in the repo and adds
the release notes for the v1.0.0 and v1.0.1 releases to it.

This release notes structure ensures that the release notes are not lost
in the event that we ever move away from github and is consistent with
how other decred repositories (dcrd, dcrpool, vspd, dcrlnd, politeia)
handle release documentation.