feat(new proposal): improve embed image button UX
This pull request improves the embed image button in the Markdown Editor.

- The old behavior: When clicking on the image button an empty markdown
image will be inserted in the editor. The user has to copy the image url and 
define the name of the image.
- The new behavior: When clicking on the image button an upload image 
modal will appear. After selecting an image, the image url will be inserted 
to the markdown text at the current location of the cursor.