feat: Add flat mode explainer.
Flat mode is not a common functionality that people are likely to be 
familar with. Comment replies may not be viewable, such as in the 
example below, so even toggling flat mode on and off is unlikely to 
provide the user with info about what it does.So this pull request 
add a button allow user to hover on it and see the expainer of 
the flat mode button.