refactor: Clear 2FA code on error.
This diff updates the 2FA workflow to clear out the 2FA code from the
modal if an error is returned by the server. The cursor is also reset
back to the first digit so that the user can easily input a new 2FA
code.

Previously, if the server returned a 2FA error the invalid code was
not cleared out and the user was forced to manually click the first
digit of the input box if they wanted to input a new code.